requirements.txt
astropy==4.3.post1
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pyds9==1.8.1
pyerfa==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0

tensorflow~=2.5.0